pyproject.toml
[tool.poetry]
name = "simple_user_app"
version = "0.1.0"
description = "Simple FastAPI App with User Management, Tailwind CSS, Dockerized Postgres and Poetry"
authors = ["Vincent Porte <vincent@neuralia.co>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
fontawesomefree = "^6.2.0"
fastapi = "^0.87.0"
uvicorn = "^0.20.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
typer = "^0.7.0"
tortoise-orm = "^0.19.2"
asyncpg = "^0.27.0"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.16.2"
black = "^22.8.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
coverage = "^6.4.4"
pre-commit = "^2.20.0"
pylint = "^2.15.2"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
aerich = "^0.7.1"


[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "./app/db/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
export = { shell = "poetry export --without-hashes --output requirements/base.txt" }
export_dev = { shell = "poetry export --without-hashes --with dev --output requirements/dev.txt" }

[tool.isort]
profile = "black"

[tool.black]
line_length = 119
